@@ -1,34 +1,34 @@
 {
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.12.0",
-    "@sentry/cli": "^1.49.0",
-    "@sentry/electron": "^1.0.0",
+    "@fortawesome/fontawesome-free": "^5.12.1",
+    "@sentry/cli": "^1.51.1",
+    "@sentry/electron": "^1.2.1",
     "@types/electron-config": "^3.2.2",
     "@types/electron-debug": "^2.1.0",
     "@types/js-yaml": "^3.12.1",
     "@types/node": "12.7.12",
-    "@types/webpack-env": "1.14.1",
-    "@typescript-eslint/eslint-plugin": "^2.13.0",
-    "@typescript-eslint/parser": "^2.14.0",
+    "@types/webpack-env": "1.15.0",
+    "@typescript-eslint/eslint-plugin": "^2.21.0",
+    "@typescript-eslint/parser": "^2.21.0",
     "apply-loader": "2.0.0",
     "awesome-typescript-loader": "^5.0.0",
-    "core-js": "^3.6.1",
-    "cross-env": "6.0.3",
-    "css-loader": "3.4.0",
-    "electron": "^7.1.7",
-    "electron-builder": "22.1.0",
+    "core-js": "^3.6.4",
+    "cross-env": "7.0.0",
+    "css-loader": "3.4.2",
+    "electron": "^8.0.2",
+    "electron-builder": "22.3.6",
     "electron-download": "^4.1.1",
-    "electron-installer-snap": "^4.1.0",
+    "electron-installer-snap": "^5.0.0",
     "electron-notarize": "^0.1.1",
-    "electron-rebuild": "^1.8.5",
+    "electron-rebuild": "^1.9.0",
     "eslint": "^6.8.0",
-    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-import": "^2.20.1",
     "file-loader": "^5.0.2",
     "graceful-fs": "^4.2.2",
     "html-loader": "0.5.5",
     "json-loader": "0.5.7",
-    "node-abi": "^2.12.0",
-    "node-gyp": "^6.0.1",
+    "node-abi": "^2.15.0",
+    "node-gyp": "^6.1.0",
     "node-sass": "^4.13.0",
     "npmlog": "4.1.2",
     "npx": "^10.2.0",
@@ -42,98 +42,29 @@
     "shelljs": "0.8.3",
     "source-code-pro": "^2.30.2",
     "source-sans-pro": "3.6.0",
-    "style-loader": "^1.1.2",
+    "style-loader": "^1.1.3",
     "svg-inline-loader": "^0.8.0",
     "to-string-loader": "1.1.6",
-    "tslib": "^1.10.0",
-    "typedoc": "^0.15.6",
-    "typescript": "^3.7.4",
+    "tslib": "^1.11.1",
+    "typedoc": "^0.16.10",
+    "typescript": "^3.8.2",
     "url-loader": "^3.0.0",
     "val-loader": "2.1.0",
-    "webpack": "^5.0.0-beta.11",
+    "webpack": "^5.0.0-beta.14",
     "webpack-cli": "^3.3.10",
     "yaml-loader": "0.5.0"
   },
   "resolutions": {
-    "*/node-abi": "^2.8.0"
-  },
-  "build": {
-    "appId": "org.terminus",
-    "productName": "Terminus",
-    "compression": "normal",
-    "afterSign": "./build/mac/afterSignHook.js",
-    "files": [
-      "**/*",
-      "dist"
-    ],
-    "extraResources": [
-      "builtin-plugins",
-      "extras"
-    ],
-    "win": {
-      "icon": "./build/windows/icon.ico",
-      "artifactName": "terminus-${version}-setup.exe",
-      "rfc3161TimeStampServer": "http://sha256timestamp.ws.symantec.com/sha256/timestamp"
-    },
-    "nsis": {
-      "oneClick": false,
-      "artifactName": "terminus-${version}-setup.${ext}",
-      "installerIcon": "./build/windows/icon.ico"
-    },
-    "publish": [
-      {
-        "provider": "github"
-      }
-    ],
-    "portable": {
-      "artifactName": "terminus-${version}-portable.exe"
-    },
-    "mac": {
-      "category": "public.app-category.video",
-      "icon": "./build/mac/icon.icns",
-      "artifactName": "terminus-${version}-macos.${ext}",
-      "hardenedRuntime": true,
-      "entitlements": "./build/mac/entitlements.plist",
-      "entitlementsInherit": "./build/mac/entitlements.plist",
-      "extendInfo": {
-        "NSRequiresAquaSystemAppearance": false
-      }
-    },
-    "pkg": {
-      "artifactName": "terminus-${version}-macos.pkg"
-    },
-    "linux": {
-      "category": "Utilities",
-      "icon": "./build/icons",
-      "artifactName": "terminus-${version}-linux.${ext}"
-    },
-    "deb": {
-      "depends": [
-        "gconf2",
-        "gconf-service",
-        "libnotify4",
-        "libsecret-1-0",
-        "libappindicator1",
-        "libxtst6",
-        "libnss3"
-      ],
-      "afterInstall": "build/linux/after-install.tpl"
-    },
-    "rpm": {
-      "depends": [
-        "screen",
-        "gnome-python2-gnomekeyring"
-      ]
-    }
+    "*/node-abi": "^2.14.0"
   },
   "scripts": {
-    "build": "npm run build:typings && webpack --color --config app/webpack.main.config.js && webpack --color --config app/webpack.config.js && webpack --color --config terminus-core/webpack.config.js && webpack --color --config terminus-settings/webpack.config.js && webpack --color --config terminus-terminal/webpack.config.js && webpack --color --config terminus-plugin-manager/webpack.config.js && webpack --color --config terminus-community-color-schemes/webpack.config.js && webpack --color --config terminus-ssh/webpack.config.js && webpack --color --config terminus-theme-hype/webpack.config.js",
-    "build:typings": "tsc --project terminus-core/tsconfig.typings.json && tsc --project terminus-settings/tsconfig.typings.json && tsc --project terminus-terminal/tsconfig.typings.json && tsc --project terminus-plugin-manager/tsconfig.typings.json && tsc --project terminus-ssh/tsconfig.typings.json",
+    "build": "npm run build:typings && webpack --color --config app/webpack.main.config.js && webpack --color --config app/webpack.config.js && webpack --color --config terminus-core/webpack.config.js && webpack --color --config terminus-settings/webpack.config.js && webpack --color --config terminus-terminal/webpack.config.js && webpack --color --config terminus-plugin-manager/webpack.config.js && webpack --color --config terminus-community-color-schemes/webpack.config.js && webpack --color --config terminus-ssh/webpack.config.js && webpack --color --config terminus-serial/webpack.config.js && webpack --color --config terminus-theme-hype/webpack.config.js",
+    "build:typings": "node scripts/build-typings.js",
     "watch": "cross-env TERMINUS_DEV=1 webpack --progress --color --watch",
     "start": "cross-env TERMINUS_DEV=1 electron app --debug",
     "prod": "cross-env TERMINUS_DEV=1 electron app",
     "docs": "typedoc --out docs/api terminus-core/src && typedoc --out docs/api/terminal --tsconfig terminus-terminal/tsconfig.typings.json terminus-terminal/src && typedoc --out docs/api/settings --tsconfig terminus-settings/tsconfig.typings.json terminus-settings/src",
-    "lint": "eslint --ext ts */src",
+    "lint": "eslint --ext ts */src */lib",
     "postinstall": "node ./scripts/install-deps.js"
   },
   "repository": "eugeny/terminus"
