@@ -21,7 +21,8 @@ exports.builtinPlugins = [
   'terminus-community-color-schemes',
   'terminus-plugin-manager',
   'terminus-ssh',
-  'terminus-theme-hype'
+  'terminus-serial',
+  'terminus-theme-hype',
 ]
 exports.bundledModules = [
   '@angular',
