@@ -1,11 +1,13 @@
 export { BaseTabComponent, BaseTabProcess } from '../components/baseTab.component'
+export { TabHeaderComponent } from '../components/tabHeader.component'
 export { SplitTabComponent, SplitContainer } from '../components/splitTab.component'
-export { TabRecoveryProvider, RecoveredTab } from './tabRecovery'
+export { TabRecoveryProvider, RecoveredTab, RecoveryToken } from './tabRecovery'
 export { ToolbarButtonProvider, ToolbarButton } from './toolbarButtonProvider'
 export { ConfigProvider } from './configProvider'
 export { HotkeyProvider, HotkeyDescription } from './hotkeyProvider'
 export { Theme } from './theme'
 export { TabContextMenuItemProvider } from './tabContextMenuProvider'
+export { SelectorOption } from './selector'
 
 export { AppService } from '../services/app.service'
 export { ConfigService } from '../services/config.service'
@@ -18,3 +20,4 @@ export { HostAppService, Platform } from '../services/hostApp.service'
 export { ShellIntegrationService } from '../services/shellIntegration.service'
 export { ThemesService } from '../services/themes.service'
 export { TabsService } from '../services/tabs.service'
+export * from '../utils'
