@@ -12,6 +12,12 @@ export interface RecoveredTab {
     options?: any
 }
 
+export interface RecoveryToken {
+    [_: string]: any
+    type: string
+    tabColor?: string|null
+}
+
 /**
  * Extend to enable recovery for your custom tab.
  * This works in conjunction with [[getRecoveryToken()]]
@@ -34,5 +40,5 @@ export abstract class TabRecoveryProvider {
      * @returns [[RecoveredTab]] descriptor containing tab type and component inputs
      *          or `null` if this token is from a different tab type or is not supported
      */
-    abstract async recover (recoveryToken: any): Promise<RecoveredTab|null>
+    abstract async recover (recoveryToken: RecoveryToken): Promise<RecoveredTab|null>
 }
