@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { Component, Input, Optional, Inject, HostBinding, HostListener, ViewChild, ElementRef } from '@angular/core'
 import { SortableComponent } from 'ng2-dnd'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
@@ -48,12 +49,15 @@ export class TabHeaderComponent {
     }
 
     ngOnInit () {
-        if (this.hostApp.platform === Platform.macOS) {
-            this.parentDraggable.setDragHandle(this.handle.nativeElement)
-        }
         this.tab.progress$.subscribe(progress => {
             this.progress = progress
         })
+    }
+
+    ngAfterViewInit () {
+        if (this.hostApp.platform === Platform.macOS) {
+            this.parentDraggable.setDragHandle(this.handle.nativeElement)
+        }
     }
 
     showRenameTabModal (): void {
