@@ -7,7 +7,7 @@ import { Directive, AfterViewInit, ElementRef } from '@angular/core'
 export class AutofocusDirective implements AfterViewInit {
     constructor (private el: ElementRef) { }
 
-    ngAfterViewInit () {
+    ngAfterViewInit (): void {
         this.el.nativeElement.blur()
         setTimeout(() => {
             this.el.nativeElement.focus()
