@@ -8,7 +8,7 @@ import { HostAppService, Platform } from './hostApp.service'
 /* eslint-disable block-scoped-var */
 
 try {
-    var wnr = require('windows-native-registry') // eslint-disable-line @typescript-eslint/no-var-requires
+    var wnr = require('windows-native-registry') // eslint-disable-line @typescript-eslint/no-var-requires, no-var
 } catch (_) { }
 
 @Injectable({ providedIn: 'root' })
@@ -18,11 +18,18 @@ export class ShellIntegrationService {
     private automatorWorkflowsDestination: string
     private registryKeys = [
         {
-            path: 'Software\\Classes\\Directory\\Background\\shell\\Open Terminus here',
+            path: 'Software\\Classes\\Directory\\Background\\shell\\Terminus',
+            value: 'Open Terminus here',
             command: 'open "%V"',
         },
         {
-            path: 'Software\\Classes\\*\\shell\\Paste path into Terminus',
+            path: 'SOFTWARE\\Classes\\Directory\\shell\\Terminus',
+            value: 'Open Terminus here',
+            command: 'open "%V"',
+        },
+        {
+            path: 'Software\\Classes\\*\\shell\\Terminus',
+            value: 'Paste path into Terminus',
             command: 'paste "%V"',
         },
     ]
@@ -51,7 +58,7 @@ export class ShellIntegrationService {
         return true
     }
 
-    async install () {
+    async install (): Promise<void> {
         const exe: string = process.env.PORTABLE_EXECUTABLE_FILE || this.electron.app.getPath('exe')
         if (this.hostApp.platform === Platform.macOS) {
             for (const wf of this.automatorWorkflows) {
@@ -61,13 +68,21 @@ export class ShellIntegrationService {
             for (const registryKey of this.registryKeys) {
                 wnr.createRegistryKey(wnr.HK.CU, registryKey.path)
                 wnr.createRegistryKey(wnr.HK.CU, registryKey.path + '\\command')
+                wnr.setRegistryValue(wnr.HK.CU, registryKey.path, '', wnr.REG.SZ, registryKey.value)
                 wnr.setRegistryValue(wnr.HK.CU, registryKey.path, 'Icon', wnr.REG.SZ, exe)
                 wnr.setRegistryValue(wnr.HK.CU, registryKey.path + '\\command', '', wnr.REG.SZ, exe + ' ' + registryKey.command)
+            }
+
+            if(wnr.getRegistryKey(wnr.HK.CU, 'Software\\Classes\\Directory\\Background\\shell\\Open Terminus here')) {
+                wnr.deleteRegistryKey(wnr.HK.CU, 'Software\\Classes\\Directory\\Background\\shell\\Open Terminus here')
+            }
+            if(wnr.getRegistryKey(wnr.HK.CU, 'Software\\Classes\\*\\shell\\Paste path into Terminus')) {
+                wnr.deleteRegistryKey(wnr.HK.CU, 'Software\\Classes\\*\\shell\\Paste path into Terminus')
             }
         }
     }
 
-    async remove () {
+    async remove (): Promise<void> {
         if (this.hostApp.platform === Platform.macOS) {
             for (const wf of this.automatorWorkflows) {
                 await exec(`rm -rf "${this.automatorWorkflowsDestination}/${wf}"`)
