@@ -1,28 +1,52 @@
-import { Component } from '@angular/core'
+import colors from 'ansi-colors'
+import { Spinner } from 'cli-spinner'
+import { Component, Injector } from '@angular/core'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
 import { first } from 'rxjs/operators'
+import { RecoveryToken } from 'terminus-core'
 import { BaseTerminalTabComponent } from 'terminus-terminal'
 import { SSHService } from '../services/ssh.service'
 import { SSHConnection, SSHSession } from '../api'
 import { SSHPortForwardingModalComponent } from './sshPortForwardingModal.component'
+import { Subscription } from 'rxjs'
 
 /** @hidden */
 @Component({
+    selector: 'ssh-tab',
     template: BaseTerminalTabComponent.template + require<string>('./sshTab.component.pug'),
     styles: [require('./sshTab.component.scss'), ...BaseTerminalTabComponent.styles],
     animations: BaseTerminalTabComponent.animations,
 })
 export class SSHTabComponent extends BaseTerminalTabComponent {
     connection: SSHConnection
-    ssh: SSHService
     session: SSHSession
-    private ngbModal: NgbModal
+    private homeEndSubscription: Subscription
 
-    ngOnInit () {
-        this.ngbModal = this.injector.get<NgbModal>(NgbModal)
+    constructor (
+        injector: Injector,
+        public ssh: SSHService,
+        private ngbModal: NgbModal,
+    ) {
+        super(injector)
+    }
 
+    ngOnInit (): void {
         this.logger = this.log.create('terminalTab')
-        this.ssh = this.injector.get(SSHService)
+
+        this.homeEndSubscription = this.hotkeys.matchedHotkey.subscribe(hotkey => {
+            if (!this.hasFocus) {
+                return
+            }
+            switch (hotkey) {
+                case 'home':
+                    this.sendInput('\x1b[H' )
+                    break
+                case 'end':
+                    this.sendInput('\x1b[F' )
+                    break
+            }
+        })
+
         this.frontendReady$.pipe(first()).subscribe(() => {
             this.initializeSession()
         })
@@ -34,7 +58,7 @@ export class SSHTabComponent extends BaseTerminalTabComponent {
         })
     }
 
-    async initializeSession () {
+    async initializeSession (): Promise<void> {
         if (!this.connection) {
             this.logger.error('No SSH connection info supplied')
             return
@@ -42,41 +66,56 @@ export class SSHTabComponent extends BaseTerminalTabComponent {
 
         this.session = this.ssh.createSession(this.connection)
         this.session.serviceMessage$.subscribe(msg => {
-            this.write(`\r\n[SSH] ${msg}\r\n`)
+            this.write('\r\n' + colors.black.bgWhite(' SSH ') + ' ' + msg + '\r\n')
             this.session.resize(this.size.columns, this.size.rows)
         })
         this.attachSessionHandlers()
         this.write(`Connecting to ${this.connection.host}`)
-        const interval = setInterval(() => this.write('.'), 500)
+
+        const spinner = new Spinner({
+            text: 'Connecting',
+            stream: {
+                write: x => this.write(x),
+            },
+        })
+        spinner.setSpinnerString(6)
+        spinner.start()
+
         try {
             await this.ssh.connectSession(this.session, (message: string) => {
-                this.write('\r\n' + message)
+                spinner.stop(true)
+                this.write(message + '\r\n')
+                spinner.start()
             })
+            spinner.stop(true)
         } catch (e) {
-            this.write('\r\n')
-            this.write(e.message)
+            spinner.stop(true)
+            this.write(colors.black.bgRed(' X ') + ' ' + colors.red(e.message) + '\r\n')
             return
-        } finally {
-            clearInterval(interval)
-            this.write('\r\n')
         }
         await this.session.start()
         this.session.resize(this.size.columns, this.size.rows)
     }
 
-    async getRecoveryToken (): Promise<any> {
+    async getRecoveryToken (): Promise<RecoveryToken> {
         return {
             type: 'app:ssh-tab',
             connection: this.connection,
+            savedState: this.frontend?.saveState(),
         }
     }
 
-    showPortForwarding () {
+    showPortForwarding (): void {
         const modal = this.ngbModal.open(SSHPortForwardingModalComponent).componentInstance as SSHPortForwardingModalComponent
         modal.session = this.session
     }
 
-    reconnect () {
+    reconnect (): void {
         this.initializeSession()
+    }
+
+    ngOnDestroy (): void {
+        this.homeEndSubscription.unsubscribe()
+        super.ngOnDestroy()
     }
 }
