@@ -38,7 +38,10 @@ module.exports = function patchPTYModule (mod) {
     }
 
     terminal.on('data', data => {
-        buffer += data
+        if (typeof data === 'string') {
+            data = Buffer.from(data)
+        }
+        buffer = Buffer.concat([buffer, data])
         if (Date.now() - lastFlush > maxWindow) {
             // Taking too much time buffering, flush to keep things interactive
             flush()
