@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { Observable } from 'rxjs'
 import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators'
 import { exec } from 'mz/child_process'
@@ -5,10 +6,9 @@ import deepEqual from 'deep-equal'
 const fontManager = require('fontmanager-redux') // eslint-disable-line
 
 import { Component, Inject } from '@angular/core'
-import { ConfigService, HostAppService, Platform, ElectronService } from 'terminus-core'
+import { ConfigService, HostAppService, Platform, ElectronService, getCSSFontFamily } from 'terminus-core'
 import { TerminalColorSchemeProvider } from '../api/colorSchemeProvider'
 import { TerminalColorScheme } from '../api/interfaces'
-import { getCSSFontFamily } from '../utils'
 
 /** @hidden */
 @Component({
