@@ -1,12 +1,11 @@
-import slug from 'slug'
+import slugify from 'slugify'
 import { Component } from '@angular/core'
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
 import { Subscription } from 'rxjs'
-import { ConfigService, ElectronService, HostAppService, Platform } from 'terminus-core'
+import { ConfigService, ElectronService, HostAppService, Platform, WIN_BUILD_CONPTY_SUPPORTED, WIN_BUILD_CONPTY_STABLE, isWindowsBuild } from 'terminus-core'
 import { EditProfileModalComponent } from './editProfileModal.component'
 import { Shell, Profile } from '../api/interfaces'
 import { TerminalService } from '../services/terminal.service'
-import { WIN_BUILD_CONPTY_SUPPORTED, WIN_BUILD_CONPTY_STABLE, isWindowsBuild } from '../utils'
 
 /** @hidden */
 @Component({
@@ -18,7 +17,7 @@ export class ShellSettingsTabComponent {
     Platform = Platform
     isConPTYAvailable: boolean
     isConPTYStable: boolean
-    slug = slug
+    slugify = slugify
     private configSubscription: Subscription
 
     constructor (
@@ -38,19 +37,19 @@ export class ShellSettingsTabComponent {
         this.isConPTYStable = isWindowsBuild(WIN_BUILD_CONPTY_STABLE)
     }
 
-    async ngOnInit () {
+    async ngOnInit (): Promise<void> {
         this.shells = await this.terminalService.shells$.toPromise()
     }
 
-    ngOnDestroy () {
+    ngOnDestroy (): void {
         this.configSubscription.unsubscribe()
     }
 
-    async reload () {
+    async reload (): Promise<void> {
         this.profiles = await this.terminalService.getProfiles(true)
     }
 
-    pickWorkingDirectory () {
+    pickWorkingDirectory (): void {
         const shell = this.shells.find(x => x.id === this.config.store.terminal.shell)
         if (!shell) {
             return
@@ -67,7 +66,7 @@ export class ShellSettingsTabComponent {
         }
     }
 
-    newProfile (shell: Shell) {
+    newProfile (shell: Shell): void {
         const profile: Profile = {
             name: shell.name || '',
             sessionOptions: this.terminalService.optionsFromShell(shell),
@@ -77,7 +76,7 @@ export class ShellSettingsTabComponent {
         this.reload()
     }
 
-    editProfile (profile: Profile) {
+    editProfile (profile: Profile): void {
         const modal = this.ngbModal.open(EditProfileModalComponent)
         modal.componentInstance.profile = Object.assign({}, profile)
         modal.result.then(result => {
@@ -86,7 +85,7 @@ export class ShellSettingsTabComponent {
         })
     }
 
-    deleteProfile (profile: Profile) {
+    deleteProfile (profile: Profile): void {
         this.config.store.terminal.profiles = this.config.store.terminal.profiles.filter(x => x !== profile)
         this.config.save()
         this.reload()
