@@ -6,19 +6,18 @@ import * as nodePTY from 'node-pty'
 import { Observable, Subject } from 'rxjs'
 import { first } from 'rxjs/operators'
 import { Injectable } from '@angular/core'
-import { Logger, LogService, ConfigService } from 'terminus-core'
+import { Logger, LogService, ConfigService, WIN_BUILD_CONPTY_SUPPORTED, isWindowsBuild } from 'terminus-core'
 import { exec } from 'mz/child_process'
 import { SessionOptions } from '../api/interfaces'
-import { WIN_BUILD_CONPTY_SUPPORTED, isWindowsBuild } from '../utils'
 
 /* eslint-disable block-scoped-var */
 
 try {
-    var macOSNativeProcessList = require('macos-native-processlist')  // eslint-disable-line @typescript-eslint/no-var-requires
+    var macOSNativeProcessList = require('macos-native-processlist')  // eslint-disable-line @typescript-eslint/no-var-requires, no-var
 } catch { }
 
 try {
-    var windowsProcessTree = require('windows-process-tree')  // eslint-disable-line @typescript-eslint/no-var-requires
+    var windowsProcessTree = require('windows-process-tree')  // eslint-disable-line @typescript-eslint/no-var-requires, no-var
 } catch { }
 
 
@@ -53,7 +52,7 @@ export abstract class BaseSession {
     get closed$ (): Observable<void> { return this.closed }
     get destroyed$ (): Observable<void> { return this.destroyed }
 
-    emitOutput (data: Buffer) {
+    emitOutput (data: Buffer): void {
         if (!this.initialDataBufferReleased) {
             this.initialDataBuffer = Buffer.concat([this.initialDataBuffer, data])
         } else {
@@ -62,7 +61,7 @@ export abstract class BaseSession {
         }
     }
 
-    releaseInitialDataBuffer () {
+    releaseInitialDataBuffer (): void {
         this.initialDataBufferReleased = true
         this.output.next(this.initialDataBuffer.toString())
         this.binaryOutput.next(this.initialDataBuffer)
@@ -100,7 +99,7 @@ export class Session extends BaseSession {
         super()
     }
 
-    start (options: SessionOptions) {
+    start (options: SessionOptions): void {
         this.name = options.name || ''
 
         const env = {
@@ -183,31 +182,13 @@ export class Session extends BaseSession {
         this.pauseAfterExit = options.pauseAfterExit || false
     }
 
-    processOSC1337 (data: Buffer) {
-        if (data.includes(OSC1337Prefix)) {
-            const preData = data.subarray(0, data.indexOf(OSC1337Prefix))
-            let params = data.subarray(data.indexOf(OSC1337Prefix) + OSC1337Prefix.length)
-            const postData = params.subarray(params.indexOf(OSC1337Suffix) + OSC1337Suffix.length)
-            const paramString = params.subarray(0, params.indexOf(OSC1337Suffix)).toString()
-
-            if (paramString.startsWith('CurrentDir=')) {
-                this.reportedCWD = paramString.split('=')[1]
-                if (this.reportedCWD.startsWith('~')) {
-                    this.reportedCWD = os.homedir() + this.reportedCWD.substring(1)
-                }
-                data = Buffer.concat([preData, postData])
-            }
-        }
-        return data
-    }
-
-    resize (columns, rows) {
+    resize (columns: number, rows: number): void {
         if (this.pty._writable) {
             this.pty.resize(columns, rows)
         }
     }
 
-    write (data: Buffer) {
+    write (data: Buffer): void {
         if (this.open) {
             if (this.pty._writable) {
                 this.pty.write(data)
@@ -217,7 +198,7 @@ export class Session extends BaseSession {
         }
     }
 
-    kill (signal?: string) {
+    kill (signal?: string): void {
         this.pty.kill(signal)
     }
 
@@ -324,6 +305,24 @@ export class Session extends BaseSession {
             this.guessedCWD = match[0]
         }
     }
+
+    private processOSC1337 (data: Buffer) {
+        if (data.includes(OSC1337Prefix)) {
+            const preData = data.subarray(0, data.indexOf(OSC1337Prefix))
+            let params = data.subarray(data.indexOf(OSC1337Prefix) + OSC1337Prefix.length)
+            const postData = params.subarray(params.indexOf(OSC1337Suffix) + OSC1337Suffix.length)
+            const paramString = params.subarray(0, params.indexOf(OSC1337Suffix)).toString()
+
+            if (paramString.startsWith('CurrentDir=')) {
+                this.reportedCWD = paramString.split('=')[1]
+                if (this.reportedCWD.startsWith('~')) {
+                    this.reportedCWD = os.homedir() + this.reportedCWD.substring(1)
+                }
+                data = Buffer.concat([preData, postData])
+            }
+        }
+        return data
+    }
 }
 
 /** @hidden */
@@ -340,7 +339,7 @@ export class SessionsService {
         this.logger = log.create('sessions')
     }
 
-    addSession (session: BaseSession, options: SessionOptions) {
+    addSession (session: BaseSession, options: SessionOptions): BaseSession {
         this.lastID++
         options.name = `session-${this.lastID}`
         session.start(options)
