@@ -1,9 +1,7 @@
 import * as path from 'path'
 import { Injectable } from '@angular/core'
-import { ElectronService } from 'terminus-core'
+import { ElectronService, WIN_BUILD_CONPTY_SUPPORTED, isWindowsBuild } from 'terminus-core'
 import { SessionOptions } from '../api/interfaces'
-
-import { WIN_BUILD_CONPTY_SUPPORTED, isWindowsBuild } from '../utils'
 
 /** @hidden */
 @Injectable({ providedIn: 'root' })
